refactor: Created header-layout file + Segregated out header code in it
